% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryOptimizationFunctions.R
\name{estimate_coefficients}
\alias{estimate_coefficients}
\title{Estimate Coefficients for Logistic Regression}
\usage{
estimate_coefficients(X, y)
}
\arguments{
\item{X}{A numeric matrix of predictor variables.}

\item{y}{A numeric vector of binary response variables (0 or 1).}
}
\value{
A numeric vector of estimated coefficients for the logistic regression model.
}
\description{
Estimates the coefficients for a logistic regression model by minimizing the logistic loss function
using the Broyden–Fletcher–Goldfarb–Shanno (BFGS) optimization method. Initial coefficient estimates are generated
using a least-squares approach.
}
\examples{
X <- matrix(c(1, 1, 1, 1, 2, 3, 4, 5), ncol = 2) # Predictors
y <- c(0, 1, 1, 0) # Binary responses
estimate_coefficients(X, y)
}
