% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryOptimizationFunctions.R
\name{predict_class}
\alias{predict_class}
\title{Predict Binary Class Labels}
\usage{
predict_class(X, beta, cutoff = 0.5)
}
\arguments{
\item{X}{A numeric matrix of predictor variables.}

\item{beta}{A numeric vector of logistic regression coefficients.}

\item{cutoff}{A numeric value between 0 and 1 specifying the classification threshold (default is 0.5).}
}
\value{
A numeric vector of predicted class labels (0 or 1).
}
\description{
Predicts binary class labels based on logistic regression coefficients and a specified cutoff.
}
\examples{
X <- matrix(c(1, 1, 1, 1, 2, 3, 4, 5), ncol = 2) # Predictors
beta <- c(-1, 0.5) # Coefficients
predict_class(X, beta, cutoff = 0.5)
}
